fix: eslint and test
